// frontend/packages/types/api/models/PatchedMedia.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
/**
 * Serializer for media files
 */
export type PatchedMedia = {
    readonly id?: number;
    folder?: number | null;
    readonly folder_name?: string;
    image?: string;
    readonly url?: string;
    readonly public_id?: string;
    file_size?: number | null;
    file_name?: string;
    used_in_posts_count?: number;
    readonly created_at?: string;
    readonly modified_at?: string;
};